refactor: User 컴포넌트 이름변경 User이라는 컴포넌트의 이름이 겹치는 현상으로인해 직전 부모의 폴더명까지 포함하여 이름을 지정합니다.
